[feat] category render 파일 분리, category 댓글 버튼 누르면 이동
